### 2023-05-12 #5419 @jarvist
This is indeed in my vague area, so I should take it!

@NicoRenaud - any suggestions for reviewers? Are you associated with any of the Ferminet / Paulinet researchers?

Generally I'd say from skim reading the Article, it seems quite early to have a JOSS paper on this package. Have you published any scientific research with it yet?
It was a bit unclear to me what the statement of need encoded: is the key thing having the variational ansatz automatically differentiable? Or an efficient GPU implementation of QMC?
Also, what flavour of QMC do you implement? Just VMC?

Separately, on a technical note, your Article PDF is > 12MB, which we need to do something about. Probably this is due to high resolution PNGs being used as figures.
