### 2023-05-12 #5419 @NicoRenaud
@jarvist  Thanks for being an editor on this.

> @NicoRenaud - any suggestions for reviewers? Are you associated with any of the Ferminet / Paulinet researchers?

I quickly talked to the Paulinet authors but we are not associated in any way.

> Generally I'd say from skim reading the Article, it seems quite early to have a JOSS paper on this package. Have you published any scientific research with it yet?

I haven't published a scientific paper about it, the main reason being a lack of time for this project on my side. Ideally it would be great to have both a scientific and a JOSS paper but the former will come much later ...

> It was a bit unclear to me what the statement of need encoded: is the key thing having the variational ansatz automatically differentiable? Or an efficient GPU implementation of QMC?

The key idea is that we have a way of encoding well tested wave function encoded in a differentiable model, which allows to easily explore different flavor of jastrow/backflow etc ...   On top of that we have GPU implementation that is not so prevalent in the QMC world. I can make both points more clear in the statement of need.

> Also, what flavour of QMC do you implement? Just VMC?

It is only VMC and now that you are pointing that out I notice it's not mentionned. I will add it to the text.

>
> Separately, on a technical note, your Article PDF is > 12MB, which we need to do something about. Probably this is due to high resolution PNGs being used as figures.

I'll reduce the figure size and reupload everything

Thanks !
