### 2023-05-15 #5419 @NicoRenaud
@jarvist I've slightly modified the text/figures to address your initial comments.

As a potential reviewer, I would suggest [Anthony Scemama](https://scemama.github.io/) who is the main developer of QMC=CHEM  and an excellent software engineer. We have interacted in the past, trying to organize a workshop with other people, but we never worked together otherwise.

You also mentionned that you though it was quite early for a JOSS paper, I'm wondering why and how to adress this concern.

Thanks !
