### 2023-05-15 #5419 @jarvist
> You also mentionned that you though it was quite early for a JOSS paper, I'm wondering why and how to adress this concern.

It's basically this bit, under 'substantial scholarly effort':
https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort

It's much easier to justify this if the code is already used in scientific papers, already used by groups outside the originator etc. In this case, I suppose you will have to explain quite directly the unique features that mean that this code will go on to be used in the future.
