### 2023-05-17 #5472 @editorialbot
[REVIEW]: QMCTorch: a PyTorch Implementation of Real-Space Quantum Monte Carlo Simulations of Molecular Systems
**Submitting author:** @NicoRenaud (<a href="http://orcid.org/0000-0001-9589-2694">Nicolas Renaud</a>)
**Repository:** https://github.com/NLESC-JCER/QMCTorch
**Branch with paper.md** (empty if default branch): master
**Version:** v0.3.2
