### 2023-05-17 NLESC-JCER/QMCTorch#145 @scemama
Review JOSS
Hi!
Here are a few comments to help improve the paper:

- [ ] line32 : varioationally -> variationally
- [ ] line43: the $D^{\uparrow}$ and $D^{\downarrow}$ determinants need to be indexed by $n$. You can do:
- [ ] $\sum_n c_n D_n^{\uparrow} D_n^{\downarrow}$, but this would imply a non-unique set of up- or down-determinants. Otherwise you can use  $\sum_{ij} c_{ij} D_i^{\uparrow} D_j^{\downarrow}$, but I am not sure that this is the representation you are using in the code.
- [ ] lines44->64 : for vector quantities, I would use bold face. For example, I would do: `$\mathbf{q}_i = \mathbf{r}_i + \sum_{i \ne j} K_{BF} (r_{ij})\,(\mathbf{r}_i - \mathbf{r}_j)$`.
- [ ] line54: You are using $\omega$ in the backflow expression, but you use also $\omega$ in the expression of the Jastrow line 49. You should use different letter to avoid confusion.
- [ ] line67: I would use a dot instead of `\times` for matrix multiplication of AO with $W_{SCF}^T$
- [ ] line86: $\rho$ is commonly used for the one-electron density. For the $N$-electron density, some authors use $\Pi$.
- [ ] line163: In the citation of Wheeler, there is some html `<tt>`
