### 2023-05-17 NLESC-JCER/QMCTorch#145 @scemama
- [ ] In the documentation, to test the software `cd test` should be replaced by `cd tests`
