### 2023-05-17 NLESC-JCER/QMCTorch#145 @scemama
- [ ] You might want also to remove the disclaimer from the README to match with the requirements of JOSS ;-)
