### 2023-05-18 #5472 @scemama
I created an issue for my review in the repo of the code: https://github.com/NLESC-JCER/QMCTorch/issues/145
