### 2023-06-10 NLESC-JCER/QMCTorch#145 @AbdAmmar
Hello ! In addition to Anthony's comments,

It would be helpful to include further references in Line 25

- [ ]  Choo et al 2020, DOI: 10.1038/s41467-020-15724-9
- [ ] Yang et al 2020, DOI: 10.1021/acs.jctc.9b01132
- [ ] Han et al 2019, DOI: 10.1016/j.jcp.2019.108929
- [ ] Kessler et al 2021,  DOI: 10.1002/adts.202000269
- [ ] Inui et al 2021, DOI: 10.1103/PhysRevResearch.3.043126
- [ ] Schätzle et al 2021, DOI: 10.1063/5.0032836
- [ ] Lin et al 2023, DOI: 10.1016/j.jcp.2022.111765

and in the "Backflow Transformation" section, Line 52
- [ ] Schmidt et al 1981, "Title of the Paper," DOI: 10.1103/PhysRevLett.47.807

- [ ] Providing some insight into the computational cost difference between using automatic differentiation and analytical expressions for the calculation of the Laplacian of the Slater determinants would be interesting. This could help readers understand the trade-offs involved in the chosen approach.
