### 2023-08-01 #5472 @scemama
Hi @NicoRenaud !
I have posted issues in the main repository to request changes:
https://github.com/NLESC-JCER/QMCTorch/issues/145
https://github.com/NLESC-JCER/QMCTorch/issues/146
I thought you would have received a notification... Next time, I will add a message here. Sorry for the misunderstanding!
