### 2023-08-29 NLESC-JCER/QMCTorch#146 @NicoRenaud
Hey @scemama thanks a lot for the review and apologies for the time it's taking me to get on to it. Thanks for finding this bug. There was a bug in the conversion when supplying the coordinate in angs, which I didn't catch. I've fixed it now as part of this PR #147
