+++ paper/paper.bib
@@ -82,7 +82,7 @@ doi={10.1007/978-3-642-38718-0_14}
 	pages = {114801},
 	author = {William A. Wheeler and Shivesh Pathak and Kevin G. Kleiner and Shunyue Yuan and Jo{\~{a}
 }o N. B. Rodrigues and Cooper Lorsung and Kittithat Krongchon and Yueqing Chang and Yiqing Zhou and Brian Busemeyer and Kiel T. Williams and Alexander Mu{\~{n}}oz and Chun Yu Chow and Lucas K. Wagner},
-	title = {$\less$tt$\greater${PyQMC}$\less$/tt$\greater$: An all-Python real-space quantum Monte Carlo module in $\less$tt$\greater${PySCF}$\less$/tt$\greater$},
+	title = {PyQMC: An all-Python real-space quantum Monte Carlo module in PySCF,
 	journal = {The Journal of Chemical Physics}
 }
 
