+++ README.md
@@ -16,7 +16,3 @@ Clone the repository and install the code from source or use the Python package 
 
 ## Documentation 
 https://qmctorch.readthedocs.io/en/latest/intro.html
-
-
-## Disclaimer
-QMCTorch is currently under developmement and most likely won't behave as expected 
