+++ paper/paper.bib
@@ -135,6 +135,123 @@ doi={10.1007/978-3-642-38718-0_14}
   Url                      = {http://dx.doi.org/10.1002/jcc.1056}
 }
 
+@article{ANN_WF,
+author = {Yang, Peng-Jian and Sugiyama, Mahito and Tsuda, Koji and Yanai, Takeshi},
+title = {Artificial Neural Networks Applied as Molecular Wave Function Solvers},
+journal = {Journal of Chemical Theory and Computation},
+volume = {16},
+number = {6},
+pages = {3513-3529},
+year = {2020},
+doi = {10.1021/acs.jctc.9b01132},
+    note ={PMID: 32320233},
+URL = { https://doi.org/10.1021/acs.jctc.9b01132},
+eprint = {https://doi.org/10.1021/acs.jctc.9b01132}
+}
+
+@article{Lin_2023,
+	doi = {10.1016/j.jcp.2022.111765},
+	url = {https://doi.org/10.1016%2Fj.jcp.2022.111765},
+	year = 2023,
+	month = {feb},
+	publisher = {Elsevier {BV}},
+	volume = {474},
+	pages = {111765},
+	author = {Jeffmin Lin and Gil Goldshlager and Lin Lin},
+	title = {Explicitly antisymmetrized neural network layers for variational Monte Carlo simulation},
+	journal = {Journal of Computational Physics}
+}
+
+@article{10.1063/5.0032836,
+    author = {Schätzle, Z. and Hermann, J. and Noé, F.},
+    title = "{Convergence to the fixed-node limit in deep variational Monte Carlo}",
+    journal = {The Journal of Chemical Physics},
+    volume = {154},
+    number = {12},
+    pages = {124108},
+    year = {2021},
+    month = {03},
+    abstract = "{Variational quantum Monte Carlo (QMC) is an ab initio method for solving the electronic Schrödinger equation that is exact in principle, but limited by the flexibility of the available Ansätze in practice. The recently introduced deep QMC approach, specifically two deep-neural-network Ansätze PauliNet and FermiNet, allows variational QMC to reach the accuracy of diffusion QMC, but little is understood about the convergence behavior of such Ansätze. Here, we analyze how deep variational QMC approaches the fixed-node limit with increasing network size. First, we demonstrate that a deep neural network can overcome the limitations of a small basis set and reach the mean-field (MF) complete-basis-set limit. Moving to electron correlation, we then perform an extensive hyperparameter scan of a deep Jastrow factor for LiH and H4 and find that variational energies at the fixed-node limit can be obtained with a sufficiently large network. Finally, we benchmark MF and many-body Ansätze on H2O, increasing the fraction of recovered fixed-node correlation energy of single-determinant Slater–Jastrow-type Ansätze by half an order of magnitude compared to previous variational QMC results, and demonstrate that a single-determinant Slater–Jastrow-backflow version of the Ansatz overcomes the fixed-node limitations. This analysis helps understand the superb accuracy of deep variational Ansätze in comparison to the traditional trial wavefunctions at the respective level of theory and will guide future improvements of the neural-network architectures in deep QMC.}",
+    issn = {0021-9606},
+    doi = {10.1063/5.0032836},
+    url = {https://doi.org/10.1063/5.0032836},
+    eprint = {https://pubs.aip.org/aip/jcp/article-pdf/doi/10.1063/5.0032836/14009445/124108\_1\_online.pdf},
+}
+
+@article{PhysRevResearch.3.043126,
+  title = {Determinant-free fermionic wave function using feed-forward neural networks},
+  author = {Inui, Koji and Kato, Yasuyuki and Motome, Yukitoshi},
[…]
