+++ paper/paper.bib
@@ -162,7 +162,7 @@ eprint = {https://doi.org/10.1021/acs.jctc.9b01132}
 	journal = {Journal of Computational Physics}
 }
 
-@article{10.1063/5.0032836,
+@article{fixed_node,
     author = {Schätzle, Z. and Hermann, J. and Noé, F.},
     title = "{Convergence to the fixed-node limit in deep variational Monte Carlo}",
     journal = {The Journal of Chemical Physics},
@@ -178,7 +178,7 @@ eprint = {https://doi.org/10.1021/acs.jctc.9b01132}
     eprint = {https://pubs.aip.org/aip/jcp/article-pdf/doi/10.1063/5.0032836/14009445/124108\_1\_online.pdf},
 }
 
-@article{PhysRevResearch.3.043126,
+@article{detfree_nn,
   title = {Determinant-free fermionic wave function using feed-forward neural networks},
   author = {Inui, Koji and Kato, Yasuyuki and Motome, Yukitoshi},
   journal = {Phys. Rev. Res.},
