### 2023-09-07 NLESC-JCER/QMCTorch#145 @scemama
> @scemama Thanks a lot for all the corrections, I've implemented all of them in #148 For the determinants I've used
> ∑
> �
> �
> �
> �
> �
> ↑
> �
> �
> ↓
>  as it is what I have implemented in the code. I could implement
> ∑
> �
> �
> �
> �
> �
> �
> �
> ↑
> �
> �
> ↓
> . Would you recommend that ?

Yes, if you have large CI expansions! You would only need to compute the unique set of $D_i^\uparrow$ and $D_j^\downarrow$ and thir gradients and Laplacian, and then you can recombine everything to get the wave function. This his how we obtain a $\sqrt{N}$ scaling for near-FCI wave functions. You can have a look at this presentation: http://irpf90.ups-tlse.fr/files/pacifichem.pdf or this paper: https://arxiv.org/abs/1510.00730
