### 2023-09-07 NLESC-JCER/QMCTorch#149 @scemama
Dead link to the doc
Hi, this link in the readme seems dead:
https://qmctorch.readthedocs.io/en/latest/intro.html
