### 2023-10-02 #5472 @scemama
@arfon @jarvist -https://github.com/NLESC-JCER/QMCTorch/issues/145 and https://github.com/NLESC-JCER/QMCTorch/issues/146  were addressed, but there are still some other bugs to fix, in https://github.com/NLESC-JCER/QMCTorch/issues/149 and https://github.com/NLESC-JCER/QMCTorch/issues/150
