### 2023-10-25 9938815
check for jastow before resetting params
