+++ qmctorch/scf/calculator/pyscf.py
@@ -51,6 +51,10 @@ class CalculatorPySCF(CalculatorBase):
             rhf {pyscf.scf} -- scf object
         """
 
+        # sphereical quantum nummbers
+        mvalues = {0: [0], 1: [-1,0,1], 2: [-2,-1,0,1,2]}
+
+        # cartesian quantum numbers
         kx = {0: [0], 1: [1, 0, 0], 2: [2, 1, 1, 0, 0, 0]}
         ky = {0: [0], 1: [0, 1, 0], 2: [0, 1, 0, 2, 1, 0]}
         kz = {0: [0], 1: [0, 0, 1], 2: [0, 0, 1, 0, 1, 2]}
@@ -58,7 +62,10 @@ class CalculatorPySCF(CalculatorBase):
         basis = SimpleNamespace()
         basis.TotalEnergy = rhf.e_tot
         basis.radial_type = 'gto_pure'
-        basis.harmonics_type = 'cart'
+        if self.basis_name.startswith('cc-'):
+            basis.harmonics_type = 'cart'
+        else:
+            basis.harmonics_type = 'cart'
 
         # number of AO / MO
         # can be different if d or f orbs are present
@@ -72,23 +79,23 @@ class CalculatorPySCF(CalculatorBase):
         # init bas properties
         bas_coeff, bas_exp = [], []
         index_ctr = []
-        bas_n, bas_l = [], []
+        bas_n, bas_m, bas_l = [], [], []
         bas_zeta = []
         bas_kx, bas_ky, bas_kz = [], [], []
-        bas_n = []
         bas_n_ori = self.get_bas_n(mol)
 
         iao = 0
+        ishell = 0
         for ibas in range(mol.nbas):
 
-            # number of zeta function per bas
-            nzeta = mol.bas_nctr(ibas)
+            # number of contracted gto per shell
+            nctr = mol.bas_nctr(ibas)
 
-            # number of contracted gaussian in that bas
-            nctr = mol.bas_nprim(ibas)
+            # number of primitive gaussian in that shell
+            nprim = mol.bas_nprim(ibas)
 
-            # number of ao from that bas <= ?
-            mult = mol.bas_len_cart(ibas) 
+            # number of cartesian component of that bas ?
+            ncart_comp = mol.bas_len_cart(ibas) 
 
             # quantum numbers
             n = bas_n_ori[ibas]
@@ -98,46 +105,58 @@ class CalculatorPySCF(CalculatorBase):
             coeffs = mol.bas_ctr_coeff(ibas)
             exps =  mol.bas_exp(ibas)
 
-            # deal with the multiple zeta
-            if coeffs.shape != (nctr, nzeta):
+            # deal with  multiple zeta
+            if coeffs.shape != (nprim, nctr):
                 raise ValueError('Contraction coefficients issue')
-            
-            if exps.shape != (nctr, nzeta):
-                exps = exps[:,np.newaxis]
-                exps = np.tile(exps,nzeta)
-                nctr *= nzeta
 
-            # coeffs/exp
-            bas_coeff += coeffs.flatten().tolist() * mult
-            bas_exp += exps.flatten().tolist() * mult 
+            # if nctr > 1:
+            #     coeffs /= np.array(range(1,nctr+1))
+            #     exps = exps[:,np.newaxis]
+            #     exps = np.tile(exps,nctr)
+            #     nprim *= nctr
+                        
+            ictr = 0
+            while ictr < nctr:
 
+                n = bas_n_ori[ishell]
+                coeffs_ictr = coeffs[:,ictr] / (ictr+1) 
 
-            # get quantum numbers per bas
-            bas_n += [n] * nctr * mult
[…]
