### 2023-10-25 5c27746
clean up get basis
