+++ .github/workflows/build.yml
@@ -28,11 +28,11 @@ jobs:
     - run: which python
     # - run: conda install -c conda-forge openmpi
     - run: conda install mpi4py h5py pytorch torchvision cpuonly -c pytorch -c conda-forge
-    # - run: pip install horovod==0.27.0
     - run: conda install -c conda-forge libstdcxx-ng=12 
+    - run: pip install horovod
 
     - name: Install the package
-      run: pip install .[test,hpc]
+      run: pip install .[test]
       env:
         CONDA_PREFIX: /usr/share/miniconda
 
