### 2023-10-25 NLESC-JCER/QMCTorch#149 @NicoRenaud
should be fixed now
