+++ .github/workflows/build.yml
@@ -30,10 +30,11 @@ jobs:
     - run: conda install mpi4py h5py pytorch torchvision cpuonly -c pytorch -c conda-forge
     - run: conda install -c conda-forge libstdcxx-ng=12 
     - run: conda install -c anaconda gxx_linux-64
-    - run: pip install --use-pep517 horovod
+    - run: pip3 install packaging
+    - run: pip3 install --use-pep517 horovod
 
     - name: Install the package
-      run: pip install .[test]
+      run: pip3 install .[test]
       env:
         CONDA_PREFIX: /usr/share/miniconda
 
