+++ .github/workflows/build.yml
@@ -31,7 +31,8 @@ jobs:
     - run: conda install -c conda-forge libstdcxx-ng=12 
     - run: conda install -c anaconda gxx_linux-64
     - run: pip3 install packaging
-    - run: pip3 install --use-pep517 horovod
+    - run: pip3 install setuptools==66.0.0
+    - run: pip3 install horovod
 
     - name: Install the package
       run: pip3 install .[test]
