+++ .github/workflows/build.yml
@@ -31,12 +31,14 @@ jobs:
     - run: conda install mpi4py h5py pytorch torchvision cpuonly -c pytorch -c conda-forge
     - run: conda install -c conda-forge libstdcxx-ng=12 
     - run: conda install -c anaconda gxx_linux-64
-    - run: pip3 install packaging
-    - run: pip3 install setuptools==66.0.0
-    - run: pip3 install horovod
+    - run: python -m pip install --upgrade pip
+    - run: python -m pip install setuptools weel 
+    - run: python -m pip install packaging
+    # - run: pip3 install setuptools==66.0.0
+    - run: python -m pip install horovod
 
     - name: Install the package
-      run: pip3 install .[test]
+      run: python -m pip install .[test]
       env:
         CONDA_PREFIX: /usr/share/miniconda
 
