+++ .github/workflows/build.yml
@@ -29,10 +29,10 @@ jobs:
     # - run: conda install -c conda-forge openmpi
     - run: conda install -c anaconda cmake
     - run: conda install mpi4py h5py pytorch torchvision cpuonly -c pytorch -c conda-forge
-    - run: conda install -c conda-forge libstdcxx-ng=12 
+    # - run: conda install -c conda-forge libstdcxx-ng=12
     - run: conda install -c anaconda gxx_linux-64
     - run: python -m pip install --upgrade pip
-    - run: python -m pip install setuptools weel 
+    - run: python -m pip install setuptools wheel 
     - run: python -m pip install packaging
     # - run: pip3 install setuptools==66.0.0
     - run: python -m pip install horovod
