+++ .github/workflows/build.yml
@@ -29,7 +29,7 @@ jobs:
     # - run: conda install -c conda-forge openmpi
     - run: conda install -c anaconda cmake
     - run: conda install mpi4py h5py pytorch torchvision cpuonly -c pytorch -c conda-forge
-    # - run: conda install -c conda-forge libstdcxx-ng=12
+    - run: conda install -c conda-forge libstdcxx-ng=12
     - run: conda install -c anaconda gxx_linux-64
     - run: python -m pip install --upgrade pip
     - run: python -m pip install setuptools wheel 
