+++ .github/workflows/build.yml
@@ -24,12 +24,14 @@ jobs:
         update-conda: true
         python-version: ${{ matrix.version }}
         conda-channels: anaconda
+    - run: sudo apt update
+    - run: sudo apt install build-essential
     - run: conda --version
     - run: which python
     # - run: conda install -c conda-forge openmpi
     - run: conda install -c anaconda cmake
     - run: conda install mpi4py h5py pytorch torchvision cpuonly -c pytorch -c conda-forge
-    - run: conda install -c conda-forge libstdcxx-ng=12
+    - run: conda install -c conda-forge libstdcxx-ng
     - run: conda install -c anaconda gxx_linux-64
     - run: python -m pip install --upgrade pip
     - run: python -m pip install setuptools wheel 
