+++ .github/workflows/build.yml
@@ -32,7 +32,8 @@ jobs:
     - name: Install conda packages
       run: |
         conda install -c anaconda cmake
-        conda install mpi4py h5py pytorch torchvision cpuonly -c pytorch -c conda-forge
+    #    conda install mpi4py h5py pytorch torchvision cpuonly -c pytorch -c conda-forge
+        conda install mpi4py h5py -c conda-forge
         conda install -c conda-forge libstdcxx-ng
         conda install -c anaconda gxx_linux-64
     - name: Instal pip packages
@@ -45,7 +46,9 @@ jobs:
         HOROVOD_WITHOUT_TENSORFLOW: 1
         HOROVOD_WITHOUT_MXNET: 1
         HOROVOD_WITHOUT_GLOO: 1
-      run: python -m pip install horovod
+      run: |
+        python -m pip torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
+        python -m pip install horovod
 
     - name: Install the package
       run: python -m pip install .[test]
