### 2023-10-26 95609a2
install pytorch from pip
