+++ .github/workflows/build.yml
@@ -49,7 +49,7 @@ jobs:
     #     python -m pip install horovod
 
     - name: Install the package
-      run: python -m pip install .[hpc, test]
+      run: python -m pip install .[test,hpc]
       env:
         CONDA_PREFIX: /usr/share/miniconda
 
