+++ .github/workflows/build.yml
@@ -34,29 +34,16 @@ jobs:
         conda install mpi4py h5py pytorch==2.0.0 torchvision==0.15.0 cpuonly -c pytorch -c conda-forge
         conda install -c conda-forge libstdcxx-ng
         conda install -c anaconda gxx_linux-64
-    # - name: Instal pip packages
-    #   run: |
-    #     python -m pip install --upgrade pip
-    #     python -m pip install packaging
-    #     python -m pip install setuptools wheel
-    # - name: Install horovod
-    #   env:
-    #     HOROVOD_WITHOUT_TENSORFLOW: 1
-    #     HOROVOD_WITHOUT_MXNET: 1
-    #     HOROVOD_WITHOUT_GLOO: 1
-    #   run: |
-    #     python -m pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
-    #     python -m pip install horovod
 
     - name: Install the package
       run: python -m pip install .[test,hpc]
       env:
         CONDA_PREFIX: /usr/share/miniconda
 
-    # - name: Test with multithreading
-    #   env:
-    #     CONDA_PREFIX: /usr/share/miniconda
-    #   run: mpirun -np 2 coverage run -m pytest tests_hvd
+    - name: Test with multithreading
+      env:
+        CONDA_PREFIX: /usr/share/miniconda
+      run: mpirun -np 2 coverage run -m pytest tests_hvd
 
     - name: Test with single thread
       env:
