### 2023-10-27 NLESC-JCER/QMCTorch#150 @NicoRenaud
@scemama I've finally found time to fix that bug, thanks for flagging it. it should be working now. However i want to refactor the way I read orbital info from pyscf and I also want to support f orbitals that are currently not supported. Thanks !
