+++ CHANGELOG.rst
@@ -12,7 +12,15 @@ Change
 * Freeze Pytorch version (#151)
 * Fix read the doc (#150)
 
+0.3.0 [Released]
+********************
 
+* Support ADF2023 (#142)
+* JOSS paper (#141)
+* Fix bug in converting distance unit (#148)
+* Hamiltonian Monte Carlo (#117)
+* Jastrow kernels (#112)
+* Backflow transformation (#115)
 
 0.1.1 [Unreleased]
 ******************
