### 2023-10-28 #5472 @NicoRenaud
Thanks @scemama ! @Kevin-Mattheus-Moerman I ve made all the changes requested by the reviewers and made a new GH and pypi release. I'm happy to do anything else ! Thanks
