### 2023-11-10 NLESC-JCER/QMCTorch#145 @tonnylou44853
Hi!

In addition to the comments above, here is a list of typos I found in the docs:
- [ ] "Jastro Kernel" should be "Jastrow Kernel" in the second line of the first paragraph in https://qmctorch.readthedocs.io/en/latest/rst/qmctorch.html.
- [ ] I believe this should be a LaTeX symbol instead of plain text: `:math: K_{ee}` in the first paragraph of https://qmctorch.readthedocs.io/en/latest/rst/qmctorch.html#two-body-jastrow-factors?
- [ ] First line in https://qmctorch.readthedocs.io/en/latest/notebooks/wfopt.html#Running-the-wave-function-optimization should be **50** steps instead of 250 steps, as it does not matches with the code below.
- [ ] https://qmctorch.readthedocs.io/en/latest/notebooks/geoopt.html#Geometry-Optimization: First paragraph, second line "firs" -> "first".
- [ ] https://qmctorch.readthedocs.io/en/latest/notebooks/geoopt.html: Cell in last line is empty.
