### 2023-11-10 #5472 @tonnylou44853
Hi! Sorry for the late response. The paper looks good to me after the corrections suggested by the other reviewers.
However, I did find a few more typos in the docs and a small bug in the code.

Here are the list of issues that I opened/commented:
https://github.com/NLESC-JCER/QMCTorch/issues/149
https://github.com/NLESC-JCER/QMCTorch/issues/145#issuecomment-1805958710
https://github.com/NLESC-JCER/QMCTorch/issues/154
