### 2023-11-11 5cfaf74
cast to int
Cast the lval to int to avoid error on macos
