### 2023-11-11 NLESC-JCER/QMCTorch#154 @NicoRenaud
Thanks @tonnylou44853  ! similar issue occur on certain install ... You're right it's safer to cast to int in the function. I'
ve made the change in a separate branch and I will merge that as soon as the test are passing
