### 2023-11-11 33a3d8f
Update qmctorch.rst
spelling mistate
