### 2023-11-11 6bf1788
Update qmctorch.rst
added math for K_{ee}
