### 2023-11-11 NLESC-JCER/QMCTorch#145 @NicoRenaud
> > @scemama Thanks a lot for all the corrections, I've implemented all of them in #148 For the determinants I've used
> > as it is what I have implemented in the code. I could implement
> > . Would you recommend that ?
>
> Yes, if you have large CI expansions! You would only need to compute the unique set of Di↑ and Dj↓ and their gradients and Laplacians. Then you can recombine everything to get the total wave function. This his how we obtain in QMC=Chem a O(Ndet) scaling for near-FCI wave functions. You can have a look at this presentation: http://irpf90.ups-tlse.fr/files/pacifichem.pdf or this paper: https://arxiv.org/abs/1510.00730 we can also have a zoom to if you want more detail :-)

Thanks for the explanation @scemama  ! That sounds like something I will have to do indeed :) I won't have too much time to develop new features on qmctorch before the end of the year. But I will pick this up in January ! I will come back to you ! Thanks !
