### 2023-11-11 NLESC-JCER/QMCTorch#145 @NicoRenaud
> Hi!
>
> In addition to the comments above, here is a list of typos I found in the docs:
>
> * [x]  "Jastro Kernel" should be "Jastrow Kernel" in the second line of the first paragraph in https://qmctorch.readthedocs.io/en/latest/rst/qmctorch.html.
> * [x]  I believe this should be a LaTeX symbol instead of plain text: `:math: K_{ee}` in the first paragraph of https://qmctorch.readthedocs.io/en/latest/rst/qmctorch.html#two-body-jastrow-factors?
> * [x]  First line in https://qmctorch.readthedocs.io/en/latest/notebooks/wfopt.html#Running-the-wave-function-optimization should be **50** steps instead of 250 steps, as it does not matches with the code below.
> * [x]  https://qmctorch.readthedocs.io/en/latest/notebooks/geoopt.html#Geometry-Optimization: First paragraph, second line "firs" -> "first".
> * [x]  https://qmctorch.readthedocs.io/en/latest/notebooks/geoopt.html: Cell in last line is empty.

Thanks @tonnylou44853  ! I've made all the change and will merge as soon as the test are ok ! :) Thanks for all the input
