### 2023-11-13 NLESC-JCER/QMCTorch#149 @NicoRenaud
Thanks @tonnylou44853 ! No apologies needed, I really appreciate the time you took to look at the code !
