### 2023-11-13 #5472 @jarvist
@NicoRenaud - a couple of the DOIs in your references are a bit garbled (I think maybe they already contain some `https://doi.org/` part of the URL?).
`Han2019` and `Kessler2021`
