### 2023-11-13 #5472 @jarvist
Great! There's a further list of tasks post-review a few messages above ^^^.

The key one is to make a release of the software; and then archive this (usually with Zenodo since you are on GitHub with the software) to get a DOI.
