### 2023-11-14 #5472 @jarvist
> openjournals/joss-papers#4777

I _think_ the XML generated looks OK - it seemed to have some issue processing the maths markup. (Or perhaps maths markup was used unnecessarily outside a maths environment?)
