### 2023-11-21 #5472 @Kevin-Mattheus-Moerman
@NicoRenaud As AEiC on this track I'll now help to process this work for acceptance in JOSS. I have check your repository, this review, the archive link, and the paper. Most seems in order. However I do have the below points that require your attention:

- [ ] The archive title should match the paper title i.e. `QMCTorch: a PyTorch Implementation of Real-Space Quantum Monte Carlo Simulations of Molecular Systems`, please edit the ZENODO title to match.
- [ ] Check spelling for:
diffentiation` ->  `differentiation`.
'heterogenous' -> 'heterogeneous'
`practitionners` ->  `practitioners`
`protoytpe` -> `prototype`
`ressources` -> `resources`
