### 2023-11-21 c7ea891
Update paper.md
more typos
