### 2023-11-21 #5472 @NicoRenaud
Thanks @Kevin-Mattheus-Moerman !

- [x]  The archive title should match the paper title i.e. QMCTorch: a PyTorch Implementation of Real-Space Quantum Monte Carlo Simulations of Molecular Systems, please edit the ZENODO title to match.
(https://zenodo.org/records/10122190)

I've also fixed all the typos you picked (sorry about that) and few extra ones
 - [x]   diffentiation -> differentiation
 - [x]  heterogenous -> heterogeneous
 - [x]  practitionners -> practitioners
 - [x]  protoytpe -> prototype
 - [x]  ressources -> resources

Let me know if can do anything else to help ! Thanks for all your work
